fixing issues realated to pull request
