--- Back End/Project2/src/main/java/com/revatue/Project2/Project2Application.java
@@ -1,14 +1,0 @@
-package com.revatue.Project2;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@SpringBootApplication
-@EnableJpaAuditing
-public class Project2Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Project2Application.class, args);
-	}
-}
