--- Back End/Project2/src/main/java/com/revatue/Project2/beans/Tag.java
@@ -1,56 +1,0 @@
-package com.revatue.Project2.beans;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.SequenceGenerator;
-import javax.persistence.Table;
-
-@Entity
-@Table(name = "TAG")
-public class Tag {
-
-	@Id
-	@SequenceGenerator(name = "NEW_TAG_SEQ", sequenceName = "NEW_TAG_SEQ", allocationSize = 1)
-	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "NEW_TAG_SEQ")
-	
-	@Column(name = "TAG_ID")
-	private int tagId;
-	@Column(name = "TAG_NAME")
-	private String tagName;
-	
-	public Tag() {
-		super();
-		}
-
-	public Tag(int tagId, String tagName) {
-		super();
-		this.tagId = tagId;
-		this.tagName = tagName;
-	}
-
-	@Override
-	public String toString() {
-		return "Tag [tagId=" + tagId + ", tagName=" + tagName + "]";
-	}
-
-	public int getTagId() {
-		return tagId;
-	}
-
-	public void setTagId(int tagId) {
-		this.tagId = tagId;
-	}
-
-	public String getTagName() {
-		return tagName;
-	}
-
-	public void setTagName(String tagName) {
-		this.tagName = tagName;
-	}
-	
-	
-}
