--- Back End/Project2/src/main/java/com/revatue/Project2/repo/CommentRepo.java
@@ -1,14 +1,0 @@
-package com.revatue.Project2.repo;
-
-import java.util.List;
-import java.util.Optional;
-
-import org.springframework.data.repository.CrudRepository;
-
-import com.revatue.Project2.beans.Comments;
-
-public interface CommentRepo extends CrudRepository<Comments, Integer>{
-	
-	 List<Comments> findByPageIdAndCommentStatus(int pageId, int commentStatus);
-
-}
