+++ BackEnd/Project2/src/main/java/com/revature/Project2Application.java
@@ -1,0 +1,14 @@
+package com.revature;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@SpringBootApplication
+@EnableJpaAuditing
+public class Project2Application {
+
+	public static void main(String[] args) {
+		SpringApplication.run(Project2Application.class, args);
+	}
+}
