+++ BackEnd/Project2/src/main/java/com/revature/beans/Page.java
@@ -1,0 +1,150 @@
+package com.revature.beans;
+
+import java.time.LocalDateTime;
+import java.util.HashSet;
+import java.util.Set;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Lob;
+import javax.persistence.OneToMany;
+import javax.persistence.SequenceGenerator;
+import javax.persistence.Table;
+
+import org.hibernate.annotations.UpdateTimestamp;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonProperty.Access;
+
+
+
+@Entity
+@Table(name = "PAGES")
+public class Page {
+
+	@Id
+	@SequenceGenerator(name = "NEW_PAGE_SEQ", sequenceName = "NEW_PAGE_SEQ", allocationSize = 1)
+	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "NEW_PAGE_SEQ")
+	@Column(name = "PAGE_ID")
+	private int pageId;
+	
+	@OneToMany(mappedBy="allComments", fetch=FetchType.LAZY)
+	private Set<Comments> pageComments = new HashSet<Comments>();
+	
+	@Column(name = "CREATED_BY_ID")
+	private int creatorId;
+	
+
+	@Column(name = "SUMMARY", columnDefinition= "CLOB")
+	@Lob
+	private String summary;
+	
+	
+	@Column(name = "BODY", columnDefinition= "CLOB")
+	@Lob
+	private String body;
+	
+	@Column(name = "TAG_ID")
+	private int tagId;
+
+
+	@Column(name = "PAGE_STATUS_ID")
+	private int pageStatus = 1;
+	
+	@Column(name = "TIMESUBMISSION")
+	@UpdateTimestamp
+	private LocalDateTime time;
+
+	public Page() {
+		super();
+	}
+
+	public Page(int pageId, int creatorId, String summary, String body, int tagId, int pageStatus, LocalDateTime time) {
+		super();
+		this.pageId = pageId;
+		this.creatorId = creatorId;
+		this.summary = summary;
+		this.body = body;
+		this.tagId = tagId;
+		this.pageStatus = pageStatus;
+		this.time = time;
+	}
+
+
+
+	public int getPageId() {
+		return pageId;
+	}
+
+	public void setPageId(int pageId) {
+		this.pageId = pageId;
+	}
+
+	public int getCreatorId() {
+		return creatorId;
+	}
+
+	public void setCreatorId(int creatorId) {
+		this.creatorId = creatorId;
+	}
+
+	public String getSummary() {
+		return summary;
+	}
+
+	public void setSummary(String summary) {
+		this.summary = summary;
+	}
+
+	public String getBody() {
+		return body;
+	}
+
+	public void setBody(String body) {
+		this.body = body;
+	}
+
+	public int getTagId() {
+		return tagId;
+	}
+
+	public void setTagId(int tagId) {
+		this.tagId = tagId;
+	}
+
+	public int getPageStatus() {
+		return pageStatus;
+	}
+
+	public void setPageStatus(int pageStatus) {
+		this.pageStatus = pageStatus;
+	}
+
+	public LocalDateTime getTime() {
+		return time;
+	}
+
+	public void setTime(LocalDateTime time) {
+		this.time = time;
+	}
+
+	public Set<Comments> getPageComments() {
+		return pageComments;
+	}
+
+	public void setPageComments(Set<Comments> pageComments) {
+		this.pageComments = pageComments;
+	}
+	
+	
+	@Override
+	public String toString() {
+		return "Page [pageId=" + pageId + ", creatorId=" + creatorId + ", summary=" + summary + ", body=" + body
+				+ ", tagId=" + tagId + ", pageStatus=" + pageStatus + ", time=" + time + "]";
+	}
+	
+}
