+++ BackEnd/Project2/src/main/java/com/revature/repo/TagRepo.java
@@ -1,0 +1,11 @@
+package com.revature.repo;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import com.revature.beans.Tag;
+
+@Repository("tagRepo")
+public interface TagRepo extends CrudRepository<Tag, Integer>{
+
+}
